some more optimizations
