+++ docs/changelog.md
@@ -1,0 +1,10 @@
+
+
+### 1.0.1
+
+ * Added react
+ * Enabled by default
+
+### 1.0.0
+
+ * Added base
