+++ docs/developers.md
@@ -1,0 +1,25 @@
+# Developers
+
+## Running locally
+
+I made this app using Electron, so you can run this locally if you want.
+
+
+#### Run locally from source code
+
+
+```
+git clone https://github.com/kevando/screenshit.git
+npm install
+electron-forge start  
+```
+
+html/css/js is in `/src` and image assets are in `/app`
+
+
+## References
+
+[Chokidar](https://ourcodeworld.com/articles/read/160/watch-files-and-directories-with-electron-framework) for watching file system.
+
+
+[Icon Converter](https://iconverticons.com/online) Export icns, export all png
