--- src/html/css/main.css
@@ -1,64 +1,0 @@
-* {
-  box-sizing: border-box;
-  font-family: 'Helvetica';
-}
-
-body {
-
-  padding: 20px 20px;
-}
-
-h2, h3 {
-  color: #333;
-}
-
-p {
-  line-height: 1.4em;
-}
-
-/* Prompt window */
-body.prompt {
-  padding: 0;
-  margin: 0;
-  background-color: #adadad;
-}
-
-body.prompt div.nav {
-  height: 28px;
-  background: blue;
-  width: 100%;
-  text-align: left;
-  padding-left: 10px;
-  margin-bottom: 10px;
-}
-.nav h1 {
-  color: #eee;
-  font-size: 14px;
-  font-weight: 500;
-  margin: 0;
-  line-height: 28px;
-}
-
-body.prompt div.icon {
-  width: 100%;
-  text-align: left;
-  padding-left: 10px;
-}
-
-body.prompt div.buttons {
-  margin-top: -5px;
-  width: 100%;
-  text-align: right;
-}
-
-body.prompt div.buttons button {
-  margin-right: 10px;
-}
-
-p.footer {
-  position: absolute;
-  bottom: 20px;
-  font-size: 12px;
-  color: #555;
-  left: 20px;
-}
