--- src/html/prompt.html
@@ -1,34 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="UTF-8">
-    <title>Frameless</title>
-    <link rel="stylesheet" href="css/main.css">
-  </head>
-  <body class='prompt'>
-    <!-- <div class='nav'><h1>New Screen Shot!</h1></div>
-
-    <div class='icon'><img src="../../app/icon/48x48.png" /></div>
-    <div class='buttons'><button type="button" id="copy-btn">Copy Image</button>
-    <button type="button" id="close-btn">Do Nothing</button></div> -->
-
-    <script>
-        // const {ipcRenderer, remote} = require('electron');
-        // var screenShotPath = remote.getCurrentWindow().screenShotPath;
-        //
-        // // Copy Clicked
-        // document.getElementById("copy-btn").addEventListener("click", function (e) {
-        //   ipcRenderer.send('copy-image', screenShotPath)
-        //   var window = remote.getCurrentWindow();
-        //   window.close();
-        // });
-        //
-        // // Close Clicked
-        // document.getElementById("close-btn").addEventListener("click", function (e) {
-        //  var window = remote.getCurrentWindow();
-        //  window.close();
-        // });
-
-    </script>
-  </body>
-</html>
